+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.1.0
+
+- feat: Add `Store.getFromPath` method.
+- feat: Add `Store.deleteFromPath` method.